refactor: export 방식 변경
