+++ pages/api/auth/[...nextauth].ts
@@ -1,4 +1,4 @@
-import { authOptions } from '@/lib/auth';
+import { authOptions } from '@/lib';
 import NextAuth from 'next-auth';
 
 export default NextAuth(authOptions);
